fix setCamera store
